feat: 调整了分组的视觉效果
